// nextjs-site/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
  id: number;
  slug: string;
  title: string;
  leadText: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  content?:
    | (
        | {
            fullWidthImage: number | Media;
            id?: string | null;
            blockName?: string | null;
            blockType: 'fullWidthImage';
          }
        | {
            heading?: string | null;
            text?: {
              root: {
                type: string;
                children: {
                  type: any;
                  version: number;
                  [k: string]: unknown;
                }[];
                direction: ('ltr' | 'rtl') | null;
                format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
                indent: number;
                version: number;
              };
              [k: string]: unknown;
            } | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'textBlock';
          }
        | {
            colSpan?: ('12' | '11' | '10' | '9' | '8' | '7' | '6' | '5' | '4' | '3') | null;
            colStart?: ('0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12') | null;
            singleMedia: number | Media;
            aspectRatio?:
              | ('auto' | 'aspect-square' | 'aspect-video' | 'aspect-[3/4]' | 'aspect-[9/16]' | 'aspect-[21/9]')
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'singleMedia';
          }
        | {
            text?: string | null;
            buttonText?: string | null;
            destination?:
              | ({
                  relationTo: 'projects';
                  value: number | Project;
                } | null)
              | ({
                  relationTo: 'caseStudies';
                  value: number | CaseStudy;
                } | null);
            /**
             * Leave empty to use the destination link
             */
            customUrl?: string | null;
            variant?: ('1' | '2') | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'CallToAction';
          }
        | {
            backgroundColor?: ('#252525' | '#e4e4e4' | 'whitesmoke') | null;
            content?:
              | (
                  | {
                      heading?: string | null;
                      text?: {
                        root: {
                          type: string;
                          children: {
                            type: any;
                            version: number;
                            [k: string]: unknown;
                          }[];
                          direction: ('ltr' | 'rtl') | null;
                          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
                          indent: number;
                          version: number;
                        };
                        [k: string]: unknown;
                      } | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'textBlock';
                    }
                  | {
                      colSpan?: ('12' | '11' | '10' | '9' | '8' | '7' | '6' | '5' | '4' | '3') | null;
                      colStart?:
                        | ('0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12')
                        | null;
                      singleMedia: number | Media;
                      aspectRatio?:
                        | (
                            | 'auto'
                            | 'aspect-square'
                            | 'aspect-video'
                            | 'aspect-[3/4]'
                            | 'aspect-[9/16]'
                            | 'aspect-[21/9]'
                          )
                        | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'singleMedia';
                    }
                )[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'recursiveComponent';
          }
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "caseStudies".
 */
export interface CaseStudy {
  id: number;
  slug: string;
  title: string;
  leadText: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  content?:
    | (
        | {
            fullWidthImage: number | Media;
            id?: string | null;
            blockName?: string | null;
            blockType: 'fullWidthImage';
          }
        | {
            heading?: string | null;
            text?: {
              root: {
                type: string;
                children: {
                  type: any;
                  version: number;
                  [k: string]: unknown;
                }[];
                direction: ('ltr' | 'rtl') | null;
                format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
                indent: number;
                version: number;
              };
              [k: string]: unknown;
            } | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'textBlock';
          }
        | {
            colSpan?: ('12' | '11' | '10' | '9' | '8' | '7' | '6' | '5' | '4' | '3') | null;
            colStart?: ('0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12') | null;
            singleMedia: number | Media;
            aspectRatio?:
              | ('auto' | 'aspect-square' | 'aspect-video' | 'aspect-[3/4]' | 'aspect-[9/16]' | 'aspect-[21/9]')
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'singleMedia';
          }
        | {
            text?: string | null;
            buttonText?: string | null;
            destination?:
              | ({
                  relationTo: 'projects';
                  value: number | Project;
                } | null)
              | ({
                  relationTo: 'caseStudies';
                  value: number | CaseStudy;
                } | null);
            /**
             * Leave empty to use the destination link
             */
            customUrl?: string | null;
            variant?: ('1' | '2') | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'CallToAction';
          }
        | {
            backgroundColor?: ('#252525' | '#e4e4e4' | 'whitesmoke') | null;
            content?:
              | (
                  | {
                      heading?: string | null;
                      text?: {
                        root: {
                          type: string;
                          children: {
                            type: any;
                            version: number;
                            [k: string]: unknown;
                          }[];
                          direction: ('ltr' | 'rtl') | null;
                          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
                          indent: number;
                          version: number;
                        };
                        [k: string]: unknown;
                      } | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'textBlock';
                    }
                  | {
                      colSpan?: ('12' | '11' | '10' | '9' | '8' | '7' | '6' | '5' | '4' | '3') | null;
                      colStart?:
                        | ('0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '10' | '11' | '12')
                        | null;
                      singleMedia: number | Media;
                      aspectRatio?:
                        | (
                            | 'auto'
                            | 'aspect-square'
                            | 'aspect-video'
                            | 'aspect-[3/4]'
                            | 'aspect-[9/16]'
                            | 'aspect-[21/9]'
                          )
                        | null;
                      id?: string | null;
                      blockName?: string | null;
                      blockType: 'singleMedia';
                    }
                )[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'recursiveComponent';
          }
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
}